requirements.txt
# Core LangChain packages with full feature set (Python 3.11.8)
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
openai>=1.50.0

# Web framework
fastapi>=0.115.0
uvicorn>=0.30.0

# Core utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
requests>=2.32.0
aiohttp>=3.10.0 